AbstractNavitiaProvider: Pass network go getLineStyles().
This allows coverages with multiple distinct transport agencies
(e.g. Australia) to meaninfully distinguish between different lines
with similar names. For example, there is often buses which are
labelled by numbers. "Bus 12" needs to be two different colours,
based on which transport network it belongs to.

In the case of Australia, it uses this to delegate to the lineStyles()
method.